refactor(web): menu icon config
