+++ web/src/components/SiderMenu/index.tsx
@@ -17,8 +17,6 @@
  * 
  * @component
  */
-
-import { UserOutlined } from '@ant-design/icons';
 import type { MenuProps } from 'antd';
 import { Menu as AntMenu, Divider, Flex, Layout } from 'antd';
 import clsx from 'clsx';
@@ -36,47 +34,6 @@ import styles from './index.module.css';
 import SubscriptionDetailModal, { type SubscriptionDetailModalRef } from './SubscriptionDetailModal';
 import SwitchSpaceModal, { type SwitchSpaceModalRef } from './SwitchSpaceModal';
 import { formatDateTime } from '@/utils/format'
-
-// Import SVG files
-// space
-import apiKeyIcon from '@/assets/images/menuNew/apiKey.svg';
-import apiKeyActiveIcon from '@/assets/images/menuNew/apiKey_active.svg';
-import applicationIcon from '@/assets/images/menuNew/application.svg';
-import applicationActiveIcon from '@/assets/images/menuNew/application_active.svg';
-import dashboardIcon from '@/assets/images/menuNew/dashboard.svg';
-import dashboardActiveIcon from '@/assets/images/menuNew/dashboard_active.svg';
-import knowledgeIcon from '@/assets/images/menuNew/knowledge.svg';
-import knowledgeActiveIcon from '@/assets/images/menuNew/knowledge_active.svg';
-import memberIcon from '@/assets/images/menuNew/member.svg';
-import memberActiveIcon from '@/assets/images/menuNew/member_active.svg';
-import memoryIcon from '@/assets/images/menuNew/memory.svg';
-import memoryActiveIcon from '@/assets/images/menuNew/memory_active.svg';
-import memoryConversationIcon from '@/assets/images/menuNew/memoryConversation.svg';
-import memoryConversationActiveIcon from '@/assets/images/menuNew/memoryConversation_active.svg';
-import ontologyIcon from '@/assets/images/menuNew/ontology.svg';
-import ontologyActiveIcon from '@/assets/images/menuNew/ontology_active.svg';
-import promptIcon from '@/assets/images/menuNew/prompt.svg';
-import promptActiveIcon from '@/assets/images/menuNew/prompt_active.svg';
-import spaceConfigIcon from '@/assets/images/menuNew/spaceConfig.svg';
-import spaceConfigActiveIcon from '@/assets/images/menuNew/spaceConfig_active.svg';
-import userMemoryIcon from '@/assets/images/menuNew/userMemory.svg';
-import userMemoryActiveIcon from '@/assets/images/menuNew/userMemory_active.svg';
-
-// manage
-import modelIcon from '@/assets/images/menuNew/model.svg';
-import modelActiveIcon from '@/assets/images/menuNew/model_active.svg';
-import pricingIcon from '@/assets/images/menuNew/pricing.svg';
-import pricingActiveIcon from '@/assets/images/menuNew/pricing_active.svg';
-import skillsIcon from '@/assets/images/menuNew/skills.svg';
-import skillsActiveIcon from '@/assets/images/menuNew/skills_active.svg';
-import spaceIcon from '@/assets/images/menuNew/space.svg';
-import spaceActiveIcon from '@/assets/images/menuNew/space_active.svg';
-import toolIcon from '@/assets/images/menuNew/tool.svg';
-import toolActiveIcon from '@/assets/images/menuNew/tool_active.svg';
-import userIcon from '@/assets/images/menuNew/user.svg';
-import userActiveIcon from '@/assets/images/menuNew/user_active.svg';
-import accountIcon from '@/assets/images/menuNew/account.svg';
-import accountActiveIcon from '@/assets/images/menuNew/account_active.svg';
 
 export interface PackagePlan {
   id: string
@@ -123,45 +80,6 @@ export interface Subscription {
   created_at: number
   updated_at: number
 }
-/** Icon path mapping table for menu items (normal and active states) */
-const iconPathMap: Record<string, string> = {
-  'dashboard': dashboardIcon,
-  'dashboardActive': dashboardActiveIcon,
-  'model': modelIcon,
-  'modelActive': modelActiveIcon,
-  'memory': memoryIcon,
-  'memoryActive': memoryActiveIcon,
-  'space': spaceIcon,
-  'spaceActive': spaceActiveIcon,
-  'user': userIcon,
-  'userActive': userActiveIcon,
-  'userMemory': userMemoryIcon,
-  'userMemoryActive': userMemoryActiveIcon,
-  'application': applicationIcon,
-  'applicationActive': applicationActiveIcon,
-  'knowledge': knowledgeIcon,
-  'knowledgeActive': knowledgeActiveIcon,
-  'memoryConversation': memoryConversationIcon,
-  'memoryConversationActive': memoryConversationActiveIcon,
-  'member': memberIcon,
-  'memberActive': memberActiveIcon,
-  'tool': toolIcon,
-  'toolActive': toolActiveIcon,
-  'apiKey': apiKeyIcon,
-  'apiKeyActive': apiKeyActiveIcon,
-  'pricing': pricingIcon,
-  'pricingActive': pricingActiveIcon,
-  'spaceConfig': spaceConfigIcon,
-  'spaceConfigActive': spaceConfigActiveIcon,
-  'ontology': ontologyIcon,
-  'ontologyActive': ontologyActiveIcon,
-  'prompt': promptIcon,
-  'promptActive': promptActiveIcon,
-  'skills': skillsIcon,
-  'skillsActive': skillsActiveIcon,
-  'account': accountIcon,
-  'accountActive': accountActiveIcon,
-};
 
 const { Sider } = Layout;
 
@@ -242,8 +160,7 @@ const Menu: FC<{
     const filteredMenus = menuList.filter(menu => menu.display);
 
     filteredMenus.forEach((menu, index) => {
-      const iconKey = selectedKeys.includes(menu.path || '') ? `${menu.code}Active` : menu.code;
-      const iconSrc = iconPathMap[iconKey as keyof typeof iconPathMap];
+      const iconKey = selectedKeys.includes(menu.path || '') ? menu.active_icon : menu.icon;
       const subs = (menu.subs || []).filter(sub => sub.display);
 
       /** Leaf node - menu item without children */
@@ -257,11 +174,9 @@ const Menu: FC<{
                 {menu.i18nKey ? t(menu.i18nKey) : menu.label}
               </span>
             ),
-            icon: iconSrc ? <img
-              src={iconSrc}
-              className="rb:w-4 rb:h-4"
-              alt={iconSrc}
-            /> : null,
+            icon: iconKey
+              ? <div className={`rb:size-4 rb:bg-cover ${iconKey}`} />
+              : null,
           });
         }
       } else {
@@ -273,11 +188,9 @@ const Menu: FC<{
           ...(menu.type === 'group' ? { type: 'group' as const } : {}),
           title: menuLabel,
           label: menuLabel,
-          icon: iconSrc ? <img
-            src={iconSrc}
-            className="rb:w-4 rb:h-4"
-            alt={iconSrc}
-          /> : <UserOutlined/>,
+          icon: iconKey
+            ? <div className={`rb:size-4 rb:bg-cover ${iconKey}`} />
+            : null,
           children,
         });
       }
