+++ web/src/store/menuData/index.ts
@@ -1,0 +1,4 @@
+import manage from './manage.json'
+import space from './space.json'
+
+export default { manage, space }
