+++ web/src/store/menuData/manage.json
@@ -1,0 +1,211 @@
+[
+  {
+    "id": 1,
+    "parent": 0,
+    "code": "workbench",
+    "label": "workbench",
+    "i18nKey": "menu.workbench",
+    "path": null,
+    "enable": true,
+    "display": true,
+    "level": 1,
+    "sort": 0,
+    "type": "group",
+    "subs": [
+      {
+        "id": 11,
+        "parent": 1,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/dashboard.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/dashboard_active.svg')]",
+        "code": "dashboard",
+        "label": "首页",
+        "i18nKey": "menu.home",
+        "path": "/index",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": []
+      },
+      {
+        "id": 12,
+        "parent": 1,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/model.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/model_active.svg')]",
+        "code": "model",
+        "label": "模型管理",
+        "i18nKey": "menu.modelManagement",
+        "path": "/model",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": []
+      },
+      {
+        "id": 13,
+        "parent": 1,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/space.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/space_active.svg')]",
+        "code": "space",
+        "label": "空间管理",
+        "i18nKey": "menu.spaceManagement",
+        "path": "/space",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": []
+      }
+    ]
+  },
+  {
+    "id": 2,
+    "parent": 0,
+    "code": "platformManagement",
+    "label": "platformManagement",
+    "i18nKey": "menu.platformManagement",
+    "path": "/",
+    "enable": true,
+    "display": true,
+    "level": 1,
+    "sort": 0,
+    "type": "group",
+    "subs": [
+      {
+        "id": 21,
+        "parent": 2,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/tool.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/tool_active.svg')]",
+        "code": "tool",
+        "label": "工具管理",
+        "i18nKey": "menu.toolManagement",
+        "path": "/tool",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": []
+      },
+      {
+        "id": 22,
+        "parent": 2,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/skills.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/skills_active.svg')]",
+        "code": "skills",
+        "label": "技能库",
+        "i18nKey": "menu.skills",
+        "path": "/skills",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": null
+      },
+      {
+        "id": 23,
+        "parent": 2,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/pricing.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/pricing_active.svg')]",
+        "code": "pricing",
+        "label": "收费管理",
+        "i18nKey": "menu.pricing",
+        "path": "/pricing",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 0,
+        "subs": [
+          {
+            "id": 61,
+            "parent": 6,
+            "code": "order",
+            "label": "订单支付",
+            "i18nKey": "menu.orderPayment",
+            "path": "/order-pay",
+            "enable": true,
+            "display": false,
+            "level": 1,
+            "sort": 0,
+            "subs": []
+          },
+          {
+            "id": 62,
+            "parent": 6,
+            "code": "orderHistory",
+            "label": "订单记录",
+            "i18nKey": "menu.orderHistory",
+            "path": "/orders",
+            "enable": true,
+            "display": false,
+            "level": 1,
+            "sort": 0,
+            "subs": []
+          }
+        ]
+      },
+      {
+        "id": 61,
+        "parent": 6,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/pricing.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/pricing_active.svg')]",
+        "code": "upgrade",
+        "label": "升级",
+        "i18nKey": "menu.upgradePackage",
+        "path": "/upgrade",
+        "enable": true,
+        "display": false,
+        "level": 1,
+        "sort": 0,
+        "subs": []
+      }
+    ]
+  },
+  {
+    "id": 3,
+    "parent": 0,
+    "code": "systemSettings",
+    "label": "systemSettings",
+    "i18nKey": "menu.systemSettings",
+    "path": null,
+    "enable": true,
+    "display": true,
+    "level": 1,
+    "sort": 0,
+    "type": "group",
+    "subs": [
+      {
+        "id": 31,
+        "parent": 3,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/user.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/user_active.svg')]",
+        "code": "user",
+        "label": "用户管理",
+        "i18nKey": "menu.userManagement",
+        "path": "/user-management",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 2,
+        "menuDesc": "用户管理",
+        "subs": null
+      },
+      {
+        "id": 32,
+        "parent": 3,
+        "icon": "rb:bg-[url('@/assets/images/menuNew/account.svg')]",
+        "active_icon": "rb:bg-[url('@/assets/images/menuNew/account_active.svg')]",
+        "code": "account",
+        "label": "账户",
+        "i18nKey": "menu.account",
+        "path": "/account",
+        "enable": true,
+        "display": true,
+        "level": 2,
+        "sort": 2,
+        "menuDesc": "账户",
+        "subs": null
+      }
+    ]
+  }
+]
